# frontend/openapi.yaml
openapi: 3.0.3
info:
  title: Claw Admin API
  description: Unified admin control plane for LLM/Skill/MCP config, SSO, organizations, users, and logs.
  version: 1.0.0
servers:
  - url: /api/v1
    description: API v1 base path
...
security:
  - BearerAuth: []
paths:
...
  /admin/users:
    get:
      tags: [admin-users]
      summary: List users
      parameters:
        - $ref: '#/components/parameters/XOrgId'
        - name: orgId
          in: query
          schema: { type: string }
        - name: role
          in: query
          schema: { type: string, enum: [owner, admin, member, viewer] }
        - name: isActive
          in: query
          schema: { type: boolean }
        - name: q
          in: query
          schema: { type: string }
        - name: cursor
          in: query
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
      responses:
        '200':
          description: Paginated users (sensitive fields masked)
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags: [admin-users]
      summary: Create user
      parameters:
        - $ref: '#/components/parameters/XOrgId'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string }
                name: { type: string }
                organizationId: { type: string }
                role: { type: string, enum: [owner, admin, member, viewer] }
      responses:
        '201':
          description: Created
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          description: Email already exists
...
  /admin/users/{id}:
    get:
      tags: [admin-users]
      summary: Get user detail (organizations and roles)
      parameters:
        - $ref: '#/components/parameters/XOrgId'
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: User with org/role list
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [admin-users]
      summary: Update user
      parameters:
        - $ref: '#/components/parameters/XOrgId'
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                avatar: { type: string }
                isActive: { type: boolean }
      responses:
        '200':
          description: Updated
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [admin-users]
      summary: Disable or delete user
      parameters:
        - $ref: '#/components/parameters/XOrgId'
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '204':
          description: Deleted or disabled
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKey:
      type: apiKey
      in: header
      name: X-Api-Key
  parameters:
    XOrgId:
      name: X-Org-Id
      in: header
      description: Tenant slug (optional; can come from JWT org_id)
      schema:
        type: string
        example: default
...
  responses:
...
    Forbidden:
      description: Forbidden (admin role required)
      content:
        application/json:
          schema:
            type: object
            properties:
              success: { type: boolean, example: false }
              error:
                type: object
                properties:
                  code: { type: string, example: FORBIDDEN }
                  message: { type: string }
...
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            type: object
            properties:
              success: { type: boolean, example: false }
              error:
                type: object
                properties:
                  code: { type: string, example: NOT_FOUND }
                  message: { type: string }